requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.6
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
oauthlib==3.1.1
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
